attendance/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import EmployeeViewSet, AttendanceViewSet

router = DefaultRouter()
router.register('employees', EmployeeViewSet)
router.register('attendances', AttendanceViewSet)

urlpatterns = [
    path('', include(router.urls)),
]
