requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
gunicorn>=21.2
psycopg2-binary>=2.9
